Move parsed_doc_from_stan to spydoc2stan to have a proper ParsedDocstring subclass
